+++ src/main/kotlin/com/example/kbocombo/auth/config/OAuthHttpConfig.kt
@@ -1,6 +1,7 @@
 package com.example.kbocombo.auth.config
 
 import com.example.kbocombo.auth.infra.KakaoHttpClient
+import com.example.kbocombo.auth.infra.KakaoAuthorizedHttpClient
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.client.RestClient
@@ -19,5 +20,16 @@ class OAuthHttpConfig {
         val factory = HttpServiceProxyFactory.builderFor(adapter)
             .build()
         return factory.createClient(KakaoHttpClient::class.java)
+    }
+
+    @Bean
+    fun kakaoAuthorizedHttpClient(): KakaoAuthorizedHttpClient {
+        val client = RestClient.builder()
+            .baseUrl("https://kapi.kakao.com")
+            .build()
+        val adapter = RestClientAdapter.create(client)
+        val factory = HttpServiceProxyFactory.builderFor(adapter)
+            .build()
+        return factory.createClient(KakaoAuthorizedHttpClient::class.java)
     }
 }
